+++ include/Nazara/Network/ENetPeer.inl
@@ -62,12 +62,12 @@ namespace Nz
 		return m_state;
 	}
 
-	inline UInt32 ENetPeer::GetTotalPacketLost() const
+	inline UInt64 ENetPeer::GetTotalPacketLost() const
 	{
 		return m_totalPacketLost;
 	}
 
-	inline UInt32 ENetPeer::GetTotalPacketSent() const
+	inline UInt64 ENetPeer::GetTotalPacketSent() const
 	{
 		return m_totalPacketSent;
 	}
